requirements.txt
langchain-groq>=0.1.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain>=0.3.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
mcp>=0.1.0
tavily-python>=0.3.0 